+++ src/savefile.h
@@ -96,6 +96,7 @@ int rd_artifacts(void);
 int rd_player_1(void);
 int rd_player_2(void);
 int rd_player_3(void);
+int rd_player_4(void);
 int rd_player(void);
 int rd_ignore(void);
 int rd_misc(void);